backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.0.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ecdsa==0.18.0
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
passlib==1.7.4
playwright==1.41.2
pluggy==1.4.0
pyasn1==0.5.1
pydantic==2.6.1
pydantic_core==2.16.2
pyee==11.0.1
PyJWT==2.8.0
pytest==8.0.1
pytest-base-url==2.1.0
pytest-playwright==0.4.4
python-jose==3.3.0
python-multipart==0.0.9
python-slugify==8.0.4
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
text-unidecode==1.3
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1